move remaining menu-related items to menu mod
